requirements.txt
requests>=2.24.0,<3.0.0
discord-webhook>=0.8.0
ConfigArgParse>=1.2.3,<2.0.0
python-dateutil>=2.8.1,<3.0.0
prompt_toolkit>=3.0.5,<4.0.0
emoji>=0.6.0
